When inferring the `dataType`, `dataset`, and `schema` for a `select_unique` query response, keen-js breaks because `parseRawData` can't find a "target property" because it's identifying some `select_unique` queries as a "Funnel" `dataType`.
PR adds a few more checks in `parseRawData` to prevent that error from
happening. Originally I wanted to refactor parseRawData to take an
analysis type from the query and to make future testing easier, but
I'm still unfamiliar with a lot of the other types of queries we can
run on explorer, so I am saving that for a later date. The WIP branch
for those changes are in `jc-parse-result-refactor`.

* look at `src/dataviz/lib/parseRawData.js`

* Go to explorer, start creating a `select_unique` query, and select a target
property that is a number type (ex: customer ids)

https://github.com/keenlabs/data-tools/issues/283